Added files via upload
more precise isFig/isApp functions:
old way:looks for "figure," "fig.," "app," etc, and assumes that it is a label, even if it is just a part of a sentence
new way: looks for "figure," "app," etc and checks to see if the next char is a number.
caveat: if the next word is a double, takes the int part and counts it as a fig/app label. don't know if it ever makes grammatical sense to write "...figure [double data]..."
future ideas: textbooks label things like figure 5.12. may want to make arrays 2D or 3D to handle stuff like this